requirements.txt
aioflask==0.4.0
docker==6.0.1
Flask==2.2.3
Flask-RESTful==0.3.9
gunicorn==20.1.0
inotify==0.2.10
PyNaCl==1.5.0
python-dotenv==1.0.0
requests==2.28.2
pytz==2022.7.1